tests/precision/requirements.txt
bitmath==1.3.3.1
Click==7.0
slack_webhook==1.0.3
ruamel.base==1.0.0
